Rename `deflate` module to `inflate`
